Instantiate second model for prediction
Load image in tf not np.
Use nested structure for image tensors.


Former-commit-id: 74c66391d41856813d6bb1cd70c1b0e263aa40c4

+++ seqtrack/pipeline.py
@@ -165,7 +165,7 @@ def load_images(example, image_size=[None, None, None], pad_value=128,
     with tf.name_scope(name) as scope:
         # Create queue to write images to.
         queue = tf.FIFOQueue(capacity=capacity,
-                             dtypes=[tf.uint8, tf.float32, tf.bool, tf.float32],
+                             dtypes=[tf.float32, tf.float32, tf.bool, tf.float32],
                              names=['images', 'labels', 'label_is_valid', 'aspect'],
                              name='image_queue')
         example = dict(example)
